refactor: TIO-264: extract helpers to keep files under 200 lines
Adding the pause/resume regression test pushed
test/integration/media_player/island_view_sync_test.dart over the
200-line per-file limit, raising the count of overlong files to 25.
Reduce it back to 23 and tighten MAX_FILE_COUNT accordingly.

- Move prepareMediaPlayerWithLoadedIsland and
  prepareMediaPlayerWithUnloadedIsland into a new
  WidgetTesterIslandViewExtension in test/utils/island_view_test_utils.dart
  (matching the existing pattern in media_player_utils.dart and
  connection_utils.dart). island_view_sync_test.dart drops 235 -> 193.
- Extract the unrelated DeleteLeadingZeros TextInputFormatter from
  lib/widgets/input/number_input_dec.dart into its own file
  lib/widgets/input/delete_leading_zeros.dart. number_input_dec.dart
  drops 215 -> 198.
- MAX_FILE_COUNT 24 -> 23. MAX_AVG_LENGTH 339 -> 345 (mechanically
  higher because the two extracted files were below the previous
  average; the per-file 200-line limit is what we actually enforce).
