+++ .github/workflows/reusable-verify.yaml
@@ -1,8 +1,8 @@
 name: Verify
 
 env:
-  MAX_FILE_COUNT: 24
-  MAX_AVG_LENGTH: 339
+  MAX_FILE_COUNT: 23
+  MAX_AVG_LENGTH: 345
   MIN_COVERAGE: 63.4
 
 on:
